refactor(sync): extract sync_engine/enhance.rs
Move the transaction enhancement pass (run_enhancement) out of
sync_engine/mod.rs into a dedicated enhance submodule.

run_enhancement drains `db.transaction_data_requests()` against
lightwalletd via GetTransaction / TransactionsInvolvingAddress and
writes the results back with decrypt_and_store_transaction /
set_transaction_status. That logic was one of the larger pieces still
sitting directly under the orchestration loop; pulling it out lets
mod.rs focus on the scan-driver control flow.

Visibility is `pub(super) use enhance::run_enhancement` — only the
orchestration loop in mod.rs calls it. All gRPC, transaction-read,
and WalletWrite imports move into enhance.rs alongside their only
consumer; the corresponding imports are dropped from mod.rs.

mod.rs is now 580 lines (from 690 after R.4). 23 tests pass.
